♻ refactor: 강한결합 service 클래스에 인터페이스를 추가해 느슨한 결합으로 리팩토링, 각 서비스 구현체에 @RequiredArgsConstructor 애너테이션 추가 BaseService 에 findEntity 메서드 삭제
