+++ Server/src/main/java/com/codestatus/domain/feed/controller/FeedController.java
@@ -1,6 +1,6 @@
 package com.codestatus.domain.feed.controller;
 
-import com.codestatus.domain.hashTag.service.HashTagService;
+import com.codestatus.domain.hashTag.service.HashTagServiceImpl;
 import com.codestatus.domain.user.mapper.UserMapper;
 import com.codestatus.global.auth.dto.PrincipalDto;
 import com.codestatus.domain.category.mapper.CategoryMapper;
@@ -10,13 +10,12 @@ import com.codestatus.domain.feed.dto.FeedPostDto;
 import com.codestatus.domain.feed.dto.FeedResponseDto;
 import com.codestatus.domain.feed.entity.Feed;
 import com.codestatus.domain.feed.mapper.FeedMapper;
-import com.codestatus.domain.feed.service.FeedService;
+import com.codestatus.domain.feed.service.FeedServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -30,9 +29,9 @@ public class FeedController {
     @Autowired
     private EntityManager entityManager;
 
-    private final FeedService feedService;
+    private final FeedServiceImpl feedServiceImpl;
 
-    private final HashTagService hashTagService;
+    private final HashTagServiceImpl hashTagServiceImpl;
 
     private final FeedMapper feedMapper;
 
@@ -41,29 +40,28 @@ public class FeedController {
     private final UserMapper userMapper;
 
 
-    public FeedController(FeedService feedService, HashTagService hashTagService, FeedMapper feedMapper, CategoryMapper categoryMapper, UserMapper userMapper) {
-        this.feedService = feedService;
-        this.hashTagService = hashTagService;
+    public FeedController(FeedServiceImpl feedServiceImpl, HashTagServiceImpl hashTagServiceImpl, FeedMapper feedMapper, CategoryMapper categoryMapper, UserMapper userMapper) {
+        this.feedServiceImpl = feedServiceImpl;
+        this.hashTagServiceImpl = hashTagServiceImpl;
         this.feedMapper = feedMapper;
         this.categoryMapper = categoryMapper;
         this.userMapper = userMapper;
     }
 
     //해당하는 카테고리에 피드 작성
-    @Transactional
     @PostMapping("/{categoryId}")
     public ResponseEntity postFeed(@PathVariable("categoryId") long categoryId,
-                                   @RequestBody FeedPostDto requstBody,
+                                   @RequestBody FeedPostDto requestBody,
                                    @AuthenticationPrincipal PrincipalDto principal) {
-        Feed feed = feedMapper.feedPostDtoToFeed(requstBody);  //dto를 엔티티로 변환
+        Feed feed = feedMapper.feedPostDtoToFeed(requestBody);  //dto를 엔티티로 변환
         feed.setCategory(categoryMapper.categoryIdToCategory(categoryId)); //피드에 카테고리 세팅
         feed.setUser((userMapper.userIdToUser(principal.getId()))); //피드에 유저 세팅
 
         entityManager.persist(feed); //영속성 컨텍스트로
 
-        List<String> hashTags = requstBody.getHashTag();// String으로 검색하기 위해 생성
-        feed.setFeedHashTags(hashTagService.createEntityByString(feed, hashTags));  //피드에 새 해쉬태그리스트 세팅
-        feedService.createEntity(feed); //피드 생성
+        List<String> hashTags = requestBody.getHashTag();// String으로 검색하기 위해 생성
+        feed.setFeedHashTags(hashTagServiceImpl.createEntityByString(feed, hashTags));  //피드에 새 해쉬태그리스트 세팅
+        feedServiceImpl.createEntity(feed); //피드 생성
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
@@ -71,7 +69,7 @@ public class FeedController {
     @GetMapping("/{categoryId}/{feedId}")
     public ResponseEntity getFeedByCategory(@PathVariable("categoryId") long categoryId,
                                             @PathVariable("feedId") long feedId) {
-        Feed feed = feedService.findFeedByCategoryAndFeedId(categoryId, feedId);
+        Feed feed = feedServiceImpl.findFeedByCategoryAndFeedId(categoryId, feedId);
 
         FeedResponseDto feedResponseDto =
                 feedMapper.feedToFeedResponseDto(feed);
@@ -82,7 +80,7 @@ public class FeedController {
     //해당 카테고리 내의 피드목록 조회
     @GetMapping("/{categoryId}")
     public ResponseEntity getFeedsByCategory(@PathVariable("categoryId") long categoryId, @RequestParam int page, @RequestParam int size) {
-        Page<Feed> pageFeeds = feedService.findAllFeedByCategory(categoryId, page-1, size);
+        Page<Feed> pageFeeds = feedServiceImpl.findAllFeedByCategory(categoryId, page-1, size);
         List<Feed> feeds = pageFeeds.getContent();
 
         return new ResponseEntity<>(
@@ -93,7 +91,7 @@ public class FeedController {
     //카테고리 구분없이 피드목록 조회
     @GetMapping
     public ResponseEntity getFeeds(@RequestParam int page, @RequestParam int size) {
-        Page<Feed> pageFeeds = feedService.findAllFeedByDeleted(page-1, size);
+        Page<Feed> pageFeeds = feedServiceImpl.findAllFeedByDeleted(page-1, size);
         List<Feed> feeds = pageFeeds.getContent();
 
         return new ResponseEntity<>(
@@ -104,7 +102,7 @@ public class FeedController {
     //일주일 내에 작성된 피드목록 중에 삭제되지 않은 피드들을 좋아요 순으로 조회
     @GetMapping("/weeklybest")
     public ResponseEntity getWeeklyBestFeeds(@RequestParam int page, @RequestParam int size) {
-        Page<Feed> pageFeeds = feedService.findWeeklyBestFeeds(page-1, size);
+        Page<Feed> pageFeeds = feedServiceImpl.findWeeklyBestFeeds(page-1, size);
         List<Feed> feeds = pageFeeds.getContent();
 
         return new ResponseEntity<>(
@@ -117,7 +115,7 @@ public class FeedController {
     public ResponseEntity getFeedsBybody(@RequestParam int page,
                                          @RequestParam int size,
                                          @RequestParam String query) {
-        Page<Feed> pageFeeds = feedService.findFeedByBody(query, page-1, size);
+        Page<Feed> pageFeeds = feedServiceImpl.findFeedByBody(query, page-1, size);
         List<Feed> feeds = pageFeeds.getContent();
 
         return new ResponseEntity<>(
@@ -131,15 +129,13 @@ public class FeedController {
                                                     @RequestParam int page,
                                          @RequestParam int size,
                                          @RequestParam String query) {
-        Page<Feed> pageFeeds = feedService.findFeedByBodyAndCategory(categoryId, query, page-1, size);
+        Page<Feed> pageFeeds = feedServiceImpl.findFeedByBodyAndCategory(categoryId, query, page-1, size);
         List<Feed> feeds = pageFeeds.getContent();
 
         return new ResponseEntity<>(
                 new MultiResponseDto<>(
                         feedMapper.feedsToFeedResponseDtos(feeds), pageFeeds), HttpStatus.OK);
     }
-
-
 
     //피드아이디로 조회하여 피드 바디 수정
     @PatchMapping("/{feedId}")
@@ -148,7 +144,7 @@ public class FeedController {
                                     @AuthenticationPrincipal PrincipalDto principal){
         Feed feed = feedMapper.feedPatchDtoToFeed(requestBody);
         feed.setFeedId(feedId);
-        feedService.updateEntity(feed, principal.getId());
+        feedServiceImpl.updateEntity(feed, principal.getId());
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
@@ -156,9 +152,7 @@ public class FeedController {
     @DeleteMapping("/{feedId}")
     public ResponseEntity deleteFeed(@PathVariable("feedId") int feedId,
                                      @AuthenticationPrincipal PrincipalDto principal) {
-        feedService.deleteEntity(feedId, principal.getId());
+        feedServiceImpl.deleteEntity(feedId, principal.getId());
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
-
-
 }
