+++ Server/src/main/java/com/codestatus/domain/status/service/StatService.java
@@ -4,6 +4,7 @@ import com.codestatus.domain.status.entity.Stat;
 import com.codestatus.domain.status.repository.StatRepository;
 import com.codestatus.global.exception.BusinessLogicException;
 import com.codestatus.global.exception.ExceptionCode;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,12 +12,9 @@ import java.util.Optional;
 
 @Transactional(readOnly = true)
 @Service
+@RequiredArgsConstructor
 public class StatService {
     private final StatRepository statRepository;
-
-    public StatService(StatRepository statRepository) {
-        this.statRepository = statRepository;
-    }
 
     public Stat findStat(long statId){
         Optional<Stat> optionalStat = statRepository.findById(statId);
